Cleanup

+++ Src/FastData/Generators/EarlyExits/Exits/UnitAtOutOfRangeEarlyExit.cs
@@ -4,12 +4,14 @@ using Genbox.FastData.Generators.Abstracts;
 
 namespace Genbox.FastData.Generators.EarlyExits.Exits;
 
-/// <summary>Rejects strings whose character at the given offset falls outside the observed [Min, Max] range using a single
-/// unsigned subtraction check: <c>UnitAt(key, offset) - Min &gt; Max - Min</c>.</summary>
+/// <summary>
+/// Rejects strings whose character at the given offset falls outside the observed [Min, Max] range using a single
+/// unsigned subtraction check: <c>UnitAt(key, offset) - Min &gt; Max - Min</c>.
+/// </summary>
 /// <remarks>Since <see cref="GeneratorFunctions.UnitAt" /> returns <see cref="uint" />, the subtraction is naturally unsigned and no cast is needed.</remarks>
 public sealed record UnitAtOutOfRangeEarlyExit(char Min, char Max, int Offset = 0) : IEarlyExit
 {
-    public ulong KeyspaceSize => (ulong)Min + (ulong)(char.MaxValue - Max);
+    public ulong KeyspaceSize => Min + (ulong)(char.MaxValue - Max);
 
     public Expression GetExpression(ParameterExpression key)
     {
